requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via
    #   langchain
    #   langchain-community
aiosignal==1.4.0
    # via aiohttp
altair==5.3.0
    # via streamlit
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.1.0
    # via
    #   gptcache
    #   streamlit
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.1.7
    # via
    #   nltk
    #   streamlit
dataclasses-json==0.5.14
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
faiss-cpu==1.7.4
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
gptcache==0.1.44
    # via langchain
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
joblib==1.5.1
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.0.142
    # via -r requirements.in
langchain-community==0.0.38
    # via -r requirements.in
langchain-core==0.1.53
    # via langchain-community
langsmith==0.1.147
    # via
    #   langchain-community
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
nltk==3.9.1
    # via textblob
numexpr==2.11.0
    # via langchain
numpy==1.26.4
    # via
    #   altair
    #   gptcache
    #   langchain
    #   langchain-community
    #   numexpr
    #   pandas
    #   pydeck
    #   streamlit
openai==1.39.0
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
orjson==3.11.1
    # via langsmith
packaging==23.2
    # via
    #   altair
    #   langchain-core
    #   marshmallow
    #   streamlit
pandas==2.3.1
    # via
    #   altair
    #   streamlit
pillow==11.3.0
    # via streamlit
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydantic==1.10.22
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   openapi-schema-pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.7.31
    # via
    #   nltk
    #   tiktoken
requests==2.32.4
    # via
    #   gptcache
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==1.4.54
    # via
    #   langchain
    #   langchain-community
streamlit==1.47.1
    # via -r requirements.in
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
textblob==0.17.1
    # via -r requirements.in
tiktoken==0.4.0
    # via -r requirements.in
toml==0.10.2
    # via streamlit
toolz==1.0.0
    # via altair
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via
    #   nltk
    #   openai
typing-extensions==4.7.1
    # via
    #   aiosignal
    #   anyio
    #   openai
    #   pydantic
    #   referencing
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
yarl==1.20.1
    # via aiohttp
